requirements_flask.txt
Flask==3.0.3
Flask-CORS==4.0.1
openai==1.66.3
requests==2.32.3
Pillow==11.1.0
numpy<2.0.0
opencv-python-headless
python-dotenv
svglib
reportlab
freetype-py
dashscope
urllib3
gunicorn==23.0.0